request adapater, controller refactor
