+++ src/assets/config.json
@@ -7,7 +7,7 @@
         "application/vnd.oasis.stix+json"
     ],
     "bind_address": "localhost",
-    "connection_string": "mongodb://repository:27017/",
+    "connection_string": "mongodb://localhost:27018/",
     "response_type": {
         "taxii": "application/vnd.oasis.taxii+json; version=2.0",
         "stix": "application/vnd.oasis.stix+json; version=2.0"
